Extract useInterval hook to separate file
